refactor(oracle): Create dedicated Oracle Free submodule
